Add vendor ID
Software that will use this vendor id: https://github.com/fredemmott/OpenKneeboard/

Spec notes that vendor IDs should be reserved, and can be used for API layers - and, if you have a vendor ID, it *should* be used for all your API layers. However, it's unclear if it's fine to publish an API layer without an assigned vendor ID:

> This usually takes the form of:
>
> An OpenXR API layer prefix string: "XR_APILAYER_"
>
> The name of the vendor developing the API layer: e.g. "LUNARG_"
>
> A short descriptive name of the API layer: e.g. "test"

If it is intended that api layers 'should' or 'must' register vendors, please approve this pull request, so I can use it in my software, which currently uses `XR_APILAYER_NOVENDOR_OpenKneeboard`

If this is not intended, perhaps the documentation should encourage the use of `NOVENDOR` or similar? This is used by several other projects, e.g. OpenXR toolkit