validation_layer_generator: do not check static array addresses
The generator creates the same validation code for a static array
of structs than for a pointer to a struct (used for dynamically
allocated arrays): it first checks if the content of the member is not
null.

Given that static array addresses are always not null, GCC (12.2.1) and
CLANG (16.0.1) emit pointer-bool-conversion warnings.